comp

+++ app/database/models.py
@@ -1,7 +1,7 @@
 """the module for creating tables"""
 from sqlalchemy import Boolean, Column, ForeignKey, Time
 from sqlalchemy.orm import relationship
-from sqlalchemy.types import Integer, String
+from sqlalchemy.types import Integer, String, BIGINT
 
 from app.database.connect import Base
 
@@ -14,7 +14,7 @@ class User(Base):
     username = Column(String(20), nullable=False)
     hashed_password = Column(String(200), nullable=False)
     # telegram_id = Column(Integer, nullable=False, unique=True)
-    telegram_id = Column(Integer, nullable=False)
+    telegram_id = Column(BIGINT, nullable=False)
     is_active = Column(Boolean, default=True)
 
     habit = relationship(
